camerad: abstract out ISP handling (#33583)
* spectra and tici

* master

* move the rest

* rm stupid indirection

* start move

* multi cam state is dead

* rest is moved

* lil more

* mv that

* lil more

* fix pc build

* we haven't done rgb for a while

* bring this stuff back

* fix mac?

* no camera in ui!

* i remember why we always cut mac

* fix mac build

---------
