+++ system/camerad/cameras/camera_common.h
@@ -8,6 +8,7 @@
 #include "common/queue.h"
 #include "common/util.h"
 
+
 const int YUV_BUFFER_COUNT = 20;
 
 enum CameraType {
@@ -39,7 +40,7 @@ typedef struct FrameMetadata {
   float processing_time;
 } FrameMetadata;
 
-struct MultiCameraState;
+class SpectraCamera;
 class CameraState;
 class ImgProc;
 
@@ -62,18 +63,13 @@ public:
 
   CameraBuf() = default;
   ~CameraBuf();
-  void init(cl_device_id device_id, cl_context context, CameraState *s, VisionIpcServer * v, int frame_cnt, VisionStreamType type);
+  void init(cl_device_id device_id, cl_context context, SpectraCamera *cam, VisionIpcServer * v, int frame_cnt, VisionStreamType type);
   bool acquire();
   void queue(size_t buf_idx);
 };
 
-void fill_frame_data(cereal::FrameData::Builder &framed, const FrameMetadata &frame_data, CameraState *c);
+void camerad_thread();
 kj::Array<uint8_t> get_raw_frame_image(const CameraBuf *b);
 float set_exposure_target(const CameraBuf *b, Rect ae_xywh, int x_skip, int y_skip);
 void publish_thumbnail(PubMaster *pm, const CameraBuf *b);
-void cameras_init(MultiCameraState *s, VisionIpcServer *v, cl_device_id device_id, cl_context ctx);
-void cameras_run(MultiCameraState *s);
-void cameras_close(MultiCameraState *s);
-void camerad_thread();
-
 int open_v4l_by_name_and_index(const char name[], int index = 0, int flags = O_RDWR | O_NONBLOCK);
